requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
generators==2020.4.27
graph-db==0.0.12
graphdb==2020.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
neo4j==5.28.1
openai==1.61.1
pqb==0.0.8
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
pytz==2025.1
requests==2.7.0
sniffio==1.3.1
starlette==0.45.3
strict-functions==2020.2.4
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
